requirements.txt
# Web Framework
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic==1.12.0
psycopg2-binary>=2.9.6
pgvector>=0.2.0

# Vector Database
weaviate-client>=3.15.0
qdrant-client>=1.1.0

# API Clients
requests==2.31.0
aiohttp==3.8.5

# Data Processing
pandas>=2.0.0
numpy>=1.24.2
beautifulsoup4==4.12.2
lxml==4.9.3
pdfminer.six==20221105

# NLP & AI
# Les versions spécifiques ci-dessous sont compatibles entre elles
huggingface_hub==0.30.1
transformers==4.41.0
sentence-transformers>=2.2.2
torch==2.3.0
scipy>=1.11.0
scikit-learn>=1.3.0
filelock==3.18.0
fsspec==2025.3.2
certifi==2025.1.31
charset-normalizer==3.4.1

# Mise à jour pour compatibilité avec NumPy 2.x
langchain>=0.1.0
langchain-community>=0.0.16
openai==0.28.0

# Workflow
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic[dotenv]>=2.0.0
email-validator>=2.0.0
# apache-airflow==2.7.0  # Commenté car crée trop de conflits

# Authentification
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=7.3.1
pytest-asyncio==0.21.1
pytest-cov>=4.1.0

# Utilities
tqdm==4.66.1
loguru>=0.7.3
schedule==1.2.0

# ETL alternatives à Airflow (plus léger)
prefect==3.3.3
textstat==0.7.5

# PDF generation
weasyprint==59.0
jinja2==3.1.6

# Added from the code block
httpx>=0.24.0
aiofiles>=23.1.0
tenacity>=8.2.2
redis>=4.5.1
gunicorn>=20.1.0
mypy>=1.2.0
flake8>=6.0.0
black>=23.3.0
